Split ObsIo source files into ObsIo and ObsFrame source files.
